orm, shani, eslint * [orm] refactor tests, use .shanirc.js * [shani] add support for disabling test files via rc * add .eslintignore file

+++ tests/glosses/orm/unit/sql/remove_column.test.js
@@ -1,23 +1,20 @@
-import Support from "../../support";
-
-
-const expectsql = Support.expectsql;
-const current = Support.sequelize;
-const sql = current.dialect.QueryGenerator;
-
 // Notice: [] will be replaced by dialect specific tick/quote character when there is not dialect specific expectation but only a default expectation
+describe("removeColumn", function () {
+    const expectsql = this.expectsql;
+    const current = this.sequelize;
+    const sql = current.dialect.QueryGenerator;
 
-describe(Support.getTestDialectTeaser("SQL"), { skip: current.dialect.name === "sqlite" }, () => {
-    describe("removeColumn", () => {
-        it("schema", () => {
-            expectsql(sql.removeColumnQuery({
-                schema: "archive",
-                tableName: "user"
-            }, "email"), {
-                mssql: "ALTER TABLE [archive].[user] DROP COLUMN [email];",
-                mysql: "ALTER TABLE `archive.user` DROP `email`;",
-                postgres: 'ALTER TABLE "archive"."user" DROP COLUMN "email";'
-            });
+    if (current.dialect.name === "sqlite") {
+        return;
+    }
+    it("schema", () => {
+        expectsql(sql.removeColumnQuery({
+            schema: "archive",
+            tableName: "user"
+        }, "email"), {
+            mssql: "ALTER TABLE [archive].[user] DROP COLUMN [email];",
+            mysql: "ALTER TABLE `archive.user` DROP `email`;",
+            postgres: 'ALTER TABLE "archive"."user" DROP COLUMN "email";'
         });
     });
 });
